Extended LLVM Values, and added basic constants
