+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "llvm-lib"
-version = "0.4.2"
+version = "0.5.0"
 authors = ["Evgeny Ukhanov <evgeny@ukhanov.org.ua>"]
 description = "LLVM library with safe and flexibility in mind, without over complexity based on LLVM-C API"
 categories = ["compilers", "development-tools", "development-tools::build-utils"]
